Refactor: switch input format
We're not yet doing anything different with it, just parsing and
verifying it's what we expected. But this change is necessary for
the next step, when we start caring about commits that were not
on the first-parent chain back from master.

Which, in turn, is exactly what we need to do to detect that a
branch is behind-only because it was already merged.